Add checkbox listener
