+++ src/index.html
@@ -12,7 +12,7 @@
 <body>
     <section>
         <div class="container">
-            <h3 class="task-box">Today's to Do <i class="fa fa-refresh"><span class="badge">20</span></i> </h3>
+            <h3 class="task-box">Today's to Do <i class="fa fa-refresh"><span class="badge">0</span></i> </h3>
             <div class="list">
                 <form>
                     <input type="text" name="addTask" id="todoInput" placeholder="Add to your list..." /><i class="addButton fa fa-reply" id="addButton"></i>
